+++ server/src/main/scala/chess/domain/CoordinateFile.scala
@@ -3,10 +3,8 @@ package chess.domain
 
 import enumeratum._
 
-/** Represents a file of a [[Coordinate]]. */
 sealed trait CoordinateFile extends EnumEntry
 
-/** Contains all [[CoordinateFile]] instances. */
 object CoordinateFile extends Enum[CoordinateFile] {
   val values: IndexedSeq[CoordinateFile] = findValues