+++ server/src/main/scala/chess/domain/CoordinateRank.scala
@@ -3,10 +3,8 @@ package chess.domain
 
 import enumeratum._
 
-/** Represents a rank of a [[Coordinate]]. */
 sealed trait CoordinateRank extends EnumEntry
 
-/** Contains all [[CoordinateRank]] instances. */
 object CoordinateRank extends Enum[CoordinateRank] {
   val values: IndexedSeq[CoordinateRank] = findValues