+++ server/src/main/scala/chess/domain/MovePatternValidator.scala
@@ -1,6 +1,8 @@
 package com.chessonline
 package chess.domain
 
+import chess.domain.CastlingType.{KingSide, QueenSide}
+import chess.domain.CoordinateFile._
 import chess.domain.CoordinateRank._
 import chess.domain.MoveValidationError.InvalidMovePattern
 import chess.domain.PieceType._
@@ -8,21 +10,14 @@ import chess.domain.Side._
 
 import cats.implicits.catsSyntaxEitherId
 
-/** Responsible for validating [[Move]] patterns. */
 trait MovePatternValidator {
 
-  /** Validates the given [[Move]] pattern using the given [[GameState]].
-    * @param move the [[Move]] to validate.
-    * @param gameState the current [[GameState]].
-    * @return [[MoveValidationError.InvalidMovePattern]] if the [[Move]] pattern is not valid, or [[MovePattern]] otherwise.
-    */
   def validate(
       move: Move,
       gameState: GameState
   ): Either[MoveValidationError, MovePattern]
 }
 
-/** The default [[MovePatternValidator]] implementation. */
 object MovePatternValidator extends MovePatternValidator {
   import MovePattern._
   type MovePatternOrError = Either[MoveValidationError, MovePattern]
@@ -31,58 +26,37 @@ object MovePatternValidator extends MovePatternValidator {
       move: Move,
       gameState: GameState
   ): MovePatternOrError = {
-
-    /** Returns the [[Side]] performed the given `move`. */
     def side: Side = move.piece.side
 
-    /** Checks whether the [[Move.to]] is empty. */
-    def destinationIsEmpty: Boolean = gameState.board(move.to).isEmpty
+    def destinationIsEmpty: Boolean = gameState.board.pieceAt(move.to).isEmpty
 
-    /** Checks whether the [[Move.to]] is taken by an enemy piece. */
-    def isAttack: Boolean = gameState.board(move.to) match {
-      case Some(attackedPiece) => side != attackedPiece.side
-      case _                   => false
-    }
+    def noPiecesBetween(
+        from: Coordinate = move.from,
+        to: Coordinate = move.to
+    ): Boolean =
+      Coordinate.allBetween(from, to).forall(gameState.board.pieceAt(_).isEmpty)
 
-    /** Checks whether there are any [[Piece]]s between the given `from` and `to`.
-      * Handles both diagonal and vertical [[Move]]s.
-      */
-    def noBarriers(forMove: Move = move): Boolean = {
-      val (fileChange, rankChange) = forMove.coordinateShift
-      val maxChange = math.max(fileChange.abs, rankChange.abs)
-
-      val initialFileIndex = CoordinateFile.indexOf(forMove.from.file)
-      val initialRankIndex = CoordinateRank.indexOf(forMove.from.rank)
-
-      val fileChangePerStep = fileChange / maxChange
-      val rankChangePerStep = rankChange / maxChange
-      val coordinatesBetween = for {
-        step <- 1 until maxChange
-      } yield Coordinate(
-        CoordinateFile.values(initialFileIndex + step * fileChangePerStep),
-        CoordinateRank.values(initialRankIndex + step * rankChangePerStep)
-      )
-
-      coordinatesBetween.forall(gameState.board(_).isEmpty)
-    }
-
-    /** Validates the [[Pawn]]'s move pattern. */
     def validateForPawn: MovePatternOrError = {
-      val isFirstMove = List(`2`, `7`).contains(move.from.rank)
-      val isEnPassantAttack = gameState.enPassantSquareOption.contains(move.to)
+      lazy val isFirstMove = List(`2`, `7`).contains(move.from.rank)
+      lazy val isEnPassantAttack =
+        gameState.enPassantSquareOption.contains(move.to)
 
-      val coordinateShift = move.coordinateShift
-      val absoluteShift = coordinateShift match {
+      val moveAsVector = move.as2DVector
+
+      // We care only about forward moves which patterns are (_, 1) | (0, 2).
+      // Move 2 squares forward for white and for black would have the following patterns respectively: (0, 2) and (0, -2).
+      // Since we care only about the pattern, we transform the rank delta taking into account which side performed the move.
+      val moveAsVectorFromSidePerspective = moveAsVector match {
         case (fileDelta, rankDelta) =>
           val newRankDelta = if (side == White) rankDelta else -rankDelta
           (fileDelta, newRankDelta)
       }
 
-      absoluteShift match {
+      moveAsVectorFromSidePerspective match {
         case (0, 1) if destinationIsEmpty => Transition.asRight
-        case (0, 2) if isFirstMove && destinationIsEmpty && noBarriers() =>
+        case (0, 2) if isFirstMove && destinationIsEmpty && noPiecesBetween() =>
           Transition.asRight
-        case (-1, 1) | (1, 1) if isAttack => Attack(move.to).asRight
+        case (-1, 1) | (1, 1) if !destinationIsEmpty => Attack(move.to).asRight
         case (-1, 1) | (1, 1) if isEnPassantAttack =>
           val toRankIndex = CoordinateRank.indexOf(move.to.rank)
           val attackedPawnRank = CoordinateRank.values(
@@ -94,49 +68,53 @@ object MovePatternValidator extends MovePatternValidator {
       }
     }
 
-//    /** Validates the move pattern for [[King]]. */
-//    def validateKingMovePattern: Either[MoveValidationError, Move] = {
-//      lazy val castlingTypeAndRookCoordinate
-//          : Option[(CastlingType, Coordinate)] =
-//        (side, move.from, move.to) match {
-//          case (White, Coordinate(E, `1`), Coordinate(C, `1`)) =>
-//            Some(WhiteQueenSide, Coordinate(A, `1`))
-//          case (White, Coordinate(E, `1`), Coordinate(G, `1`)) =>
-//            Some(WhiteKingSide, Coordinate(H, `1`))
-//          case (Black, Coordinate(E, `8`), Coordinate(C, `8`)) =>
-//            Some(BlackQueenSide, Coordinate(A, `8`))
-//          case (Black, Coordinate(E, `8`), Coordinate(G, `8`)) =>
-//            Some(BlackKingSide, Coordinate(H, `8`))
-//          case _ => None
-//        }
-//
-//      val (fileChange, rankChange) = moveDistance()
-//      (fileChange.abs, rankChange.abs) match {
-//        case (1, 0) | (0, 1) if destinationIsEmpty || isAttack =>
-//          Right(move)
-//
-//        case _ =>
-//          castlingTypeAndRookCoordinate match {
-//            case Some((castling, rookCoordinate))
-//                if gameState.castlingsAvailable.contains(
-//                  castling
-//                ) && noBarriers(to = rookCoordinate) =>
-//              Right(move)
-//            case _ => Left(InvalidMovePattern)
-//          }
-//      }
-//    }
+    def validateForKing: Either[MoveValidationError, MovePattern] = {
+      val (fileDelta, rankDelta) = move.as2DVector
+
+      lazy val castlingType: Option[CastlingType] = {
+        val castlingType = (fileDelta, rankDelta) match {
+          case (2, 0)  => Some(KingSide)
+          case (-2, 0) => Some(QueenSide)
+          case _       => None
+        }
+        val kingInitialCoordinate = side match {
+          case White => Coordinate(E, `1`)
+          case Black => Coordinate(E, `8`)
+        }
+
+        for {
+          castlingType <- castlingType
+          if move.from == kingInitialCoordinate
+
+          rookFile = castlingType match {
+            case QueenSide => A
+            case KingSide  => H
+          }
+          rookCoordinate = move.from.copy(file = rookFile)
+          if noPiecesBetween(to = rookCoordinate)
+        } yield castlingType
+      }
+
+      (fileDelta.abs, rankDelta.abs) match {
+        case (1, 0) | (0, 1) | (1, 1) =>
+          if (destinationIsEmpty) Transition.asRight
+          else Attack(move.to).asRight
+
+        case _ =>
+          (for {
+            castlingType <- castlingType
+            if gameState.castingAvailable(side, castlingType)
+          } yield Castling(castlingType)).toRight(left = InvalidMovePattern)
+      }
+    }
 
     move.piece.pieceType match {
-      case Pawn => validateForPawn
-//      case King => validateKingMovePattern
-      case _ => Left(InvalidMovePattern)
-//        case PieceType.King => ???
-//        case PieceType.Queen => ???
-//        case PieceType.Rook => ???
-//        case PieceType.Bishop => ???
-//        case PieceType.Knight => ???
+      case Pawn   => validateForPawn
+      case King   => validateForKing
+      case Queen  => ???
+      case Rook   => ???
+      case Bishop => ???
+      case Knight => ???
     }
   }
-
 }