+++ server/src/main/scala/chess/domain/PieceType.scala
@@ -3,10 +3,8 @@ package chess.domain
 
 import enumeratum._
 
-/** Represents a type of a [[Piece]] */
 sealed trait PieceType extends EnumEntry
 
-/** Contains all [[Piece]] instances. */
 object PieceType extends Enum[PieceType] {
   val values: IndexedSeq[PieceType] = findValues