+++ server/src/test/scala/chess/domain/MoveSpec.scala
@@ -8,12 +8,12 @@ class MoveSpec extends AnyFreeSpec {
   "Move" - {
     import TestData._
 
-    "coordinateShift" - {
-      "should return correct coordinate shift" in {
-        Move(whitePawn, a1, a2).coordinateShift shouldEqual (0, 1)
-        Move(whitePawn, a1, b1).coordinateShift shouldEqual (1, 0)
-        Move(whitePawn, a1, a1).coordinateShift shouldEqual (0, 0)
-        Move(whitePawn, a1, b2).coordinateShift shouldEqual (1, 1)
+    "as2DVector" - {
+      "should return correct vector representation" in {
+        Move(whitePawn, a1, a2).as2DVector shouldEqual (0, 1)
+        Move(whitePawn, a1, b1).as2DVector shouldEqual (1, 0)
+        Move(whitePawn, a1, a1).as2DVector shouldEqual (0, 0)
+        Move(whitePawn, a1, b2).as2DVector shouldEqual (1, 1)
       }
     }
   }